Create www.conf
Configure Processes

PHP-FPM create processes to handle incoming requests. If it runs out of available processes, new requests will be put in queue and the users will experience delays.

The number of processes to use will heavily depend on the traffic on your websites and on the available RAM/CPU on your server. To find optimal values check the log file /var/log/php7.0-fpm.log. It will contains warnings when the number of processes need adjustment.